feedback + spoonacular integration

+++ apps/mehungry/lib/mehungry/food.ex
@@ -241,7 +241,7 @@ defmodule Mehungry.Food do
       Repo.delete(rec_in)
     end)
 
-    if !is_nil(recipe.image_url) do
+    if !is_nil(recipe.image_url) and not String.starts_with?(recipe.image_url, "http") do
       file_name = List.last(String.split(recipe.image_url))
 
       ExAws.S3.delete_object("test-bucket-local-mehungry", file_name)
